requirements.txt
discord.py
requests
jpholiday
